Fix Gradle sync in Android Studio
This change is based on recommendations described at
https://developer.android.com/build/dependency-verification and solves a
problem with the codebase not loading properly in Android Studio.

The official documentation recommends trusting the Gradle group
entirely. I've instead narrowed it down to one specific Gradle checksum
from their GitHub release page.

It would be more secure to instead trust specific hashes of all source
and Javadoc JARs. I have not found a way to do so, so I resorted to the
official recommendation of using two `<trusted-artifacts>` rules.